// ui/src/generated/ShareNotice.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Color } from "./Color";
import type { ProjectId } from "./ProjectId";

export type ShareNotice = { "type": "ParticipantJoined", name: string, color: Color, } | { "type": "ParticipantLeft", name: string, } | { "type": "HostOffline", host_name: string, } | { "type": "HostBack", host_name: string, } | { "type": "SharingEnded", host_name: string, } | { "type": "LocalCopyKept", project: ProjectId, name: string, };